fixed links, showing only for logged in users. created middleware for admin users.

+++ larabook/routes/web.php
@@ -1,6 +1,6 @@
 <?php
-
-use App\Models\Venue;
+use App\Http\Middleware\Admin;
+use App\Http\Middleware\LoggedIn;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\VenueController;
 use App\Http\Controllers\SessionController;
@@ -13,18 +13,16 @@ Route::get('/', function () {
 //venues
 Route::get('/venue/getList', [VenueController::class, 'index']);
 Route::get('/venue/getUsersVenues', [VenueController::class, 'getUsersVenues']);
-Route::get('/venue/create', [VenueController::class, 'create']);
-Route::post('/venue/store', [VenueController::class, 'store']);
-Route::get('/venue/edit/{id}', [VenueController::class, 'edit']);
-Route::patch('/venue/{id}', [VenueController::class, 'update']);
-Route::delete('/venue/{id}', [VenueController::class, 'destroy']);
+Route::get('/venue/create', [VenueController::class, 'create'])->middleware(Admin::class);
+Route::post('/venue/store', [VenueController::class, 'store'])->middleware(Admin::class);
+Route::get('/venue/edit/{id}', [VenueController::class, 'edit'])->middleware(Admin::class);
+Route::patch('/venue/{id}', [VenueController::class, 'update'])->middleware(Admin::class);
+Route::delete('/venue/{id}', [VenueController::class, 'destroy'])->middleware(Admin::class);
 Route::get('/venue/{id}', [VenueController::class, 'show']);
 Route::get('/venueRequest/{id}', [VenueController::class, 'findRequest']);
+Route::get('/myVenues/index', [VenueController::class, 'myVenues'])->middleware(Admin::class);
 
-//my venues page
-Route::get('/myVenues/index', function () {
-    return view('myVenues.index');
-});
+
 
 Route::get('/dashboard', function () {
     return view('dashboard');
@@ -32,7 +30,7 @@ Route::get('/dashboard', function () {
 //auth
 Route::get('/register', [RegisteredUserController::class, 'create']);
 Route::post('/register', [RegisteredUserController::class, 'store']);
-Route::get('/profile', [RegisteredUserController::class, 'show']);
+Route::get('/profile', [RegisteredUserController::class, 'show'])->middleware(LoggedIn::class);
 
 Route::get('/login', [SessionController::class, 'create']);
 Route::post('/login', [SessionController::class, 'store']);
